# dist/utils/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Quizmon API",
        "description": "REST API documentation for the Quizmon quiz platform. Supports quiz creation, match hosting, AI-powered question generation, and real-time gameplay via Socket.IO.",
        "version": "1.0.0"
    },
    "host": "",
    "basePath": "/",
    "schemes": [
        "http"
    ],
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "JWT Bearer token. Format: `Bearer <token>`"
        }
    },
    "paths": {
...
        "/quiz/": {
            "post": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Create a new quiz",
                "description": "Creates a quiz for the authenticated user. Accepts an optional image file upload (multipart/form-data).",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "World Capitals Quiz"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "example": "Test your geography knowledge"
                    },
                    {
                        "name": "isPublic",
                        "in": "formData",
                        "required": false,
                        "type": "boolean",
                        "example": true
                    },
                    {
                        "name": "categoryId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "example": 2
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "required": false,
                        "type": "file",
                        "description": "Cover image for the quiz (uploaded to Cloudinary)"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Quiz created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Get all quizzes for the current user",
                "description": "Returns a list of quizzes created by the authenticated user, including creator and category info.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quizzes returned"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/quiz/{id}": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Get a single quiz with full details",
                "description": "Retrieves a quiz by ID, including all questions and their sub-type data (buttons, checkboxes, range, etc.).",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz returned successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/quiz/{id}/question": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Get all questions for a quiz",
                "description": "Returns all questions belonging to the given quiz, each with its type-specific data.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Questions returned successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/quiz/{id}/rated": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Check if current user has rated a quiz",
                "description": "Returns `{ rated: true/false }` indicating whether the authenticated user has already submitted a rating for this quiz.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating check result",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "rated": {
                                    "type": "boolean",
                                    "example": false
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/quiz/{id}/rating": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Get ratings for a quiz",
                "description": "Returns all ratings for the quiz, along with the aggregate average score and total count.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ratings returned successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "average": {
                                    "type": "number",
                                    "example": 4.2
                                },
                                "count": {
                                    "type": "integer",
                                    "example": 15
                                },
                                "ratings": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "userId": {
                                                "type": "integer"
                                            },
                                            "rating": {
                                                "type": "integer"
                                            },
                                            "text": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/category/{id}/quiz": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get public quizzes by category",
                "description": "Returns all public quizzes that belong to the specified category.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quizzes returned successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}